refactor: create ExploreMetricButton component
