+++ packages/frontend/src/features/metricsCatalog/components/MetricsTable.tsx
@@ -20,16 +20,12 @@ import {
     type UIEvent,
 } from 'react';
 import MantineIcon from '../../../components/common/MantineIcon';
-import { useExplore } from '../../../hooks/useExplore';
-import {
-    createMetricPreviewUnsavedChartVersion,
-    getExplorerUrlFromCreateSavedChartVersion,
-} from '../../../hooks/useExplorerRoute';
 import { useTracking } from '../../../providers/TrackingProvider';
 import { EventName } from '../../../types/Events';
 import { useAppDispatch, useAppSelector } from '../../sqlRunner/store/hooks';
 import { useMetricsCatalog } from '../hooks/useMetricsCatalog';
 import { setActiveMetric } from '../store/metricsCatalogSlice';
+import { ExploreMetricButton } from './ExploreMetricButton';
 
 const MetricUsageButton = ({ row }: { row: MRT_Row<CatalogField> }) => {
     const hasChartsUsage = row.original.chartUsage ?? 0 > 0;
@@ -143,62 +139,6 @@ const columns: MRT_ColumnDef<CatalogField>[] = [
         Cell: ({ row }) => <MetricUsageButton row={row} />,
     },
 ];
-
-const UseMetricButton = ({ row }: { row: MRT_Row<CatalogField> }) => {
-    const [isGeneratingPreviewUrl, setIsGeneratingPreviewUrl] = useState(false);
-    const projectUuid = useAppSelector(
-        (state) => state.metricsCatalog.projectUuid,
-    );
-    const [currentTableName, setCurrentTableName] = useState<string>();
-    const { track } = useTracking();
-    const { isFetching } = useExplore(currentTableName, {
-        onSuccess(explore) {
-            if (!!currentTableName && explore && projectUuid) {
-                setIsGeneratingPreviewUrl(true);
-                const unsavedChartVersion =
-                    createMetricPreviewUnsavedChartVersion(
-                        row.original,
-                        explore,
-                    );
-
-                const { pathname, search } =
-                    getExplorerUrlFromCreateSavedChartVersion(
-                        projectUuid,
-                        unsavedChartVersion,
-                    );
-                const url = new URL(pathname, window.location.origin);
-                url.search = new URLSearchParams(search).toString();
-
-                window.open(url.href, '_blank');
-                setIsGeneratingPreviewUrl(false);
-                setCurrentTableName(undefined);
-            }
-        },
-    });
-
-    const handleExploreClick = () => {
-        track({
-            name: EventName.METRICS_CATALOG_EXPLORE_CLICKED,
-            properties: {
-                metricName: row.original.name,
-                tableName: row.original.tableName,
-            },
-        });
-        setCurrentTableName(row.original.tableName);
-    };
-
-    return (
-        <Button
-            size="xs"
-            compact
-            variant="subtle"
-            onClick={handleExploreClick}
-            loading={isFetching || isGeneratingPreviewUrl}
-        >
-            Explore
-        </Button>
-    );
-};
 
 export const MetricsTable = () => {
     const projectUuid = useAppSelector(
@@ -341,7 +281,7 @@ export const MetricsTable = () => {
         },
         renderRowActions: ({ row }) => (
             <Box>
-                <UseMetricButton row={row} />
+                <ExploreMetricButton row={row} />
             </Box>
         ),
         enableFilterMatchHighlighting: true,
